backend/requirements.txt
# AFM Pipeline Backend Requirements
# Install with: pip install -r requirements.txt
# Or use conda: conda env create -f environment.yml

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.0.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0,<4.12
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0,<2.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Deep Learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0

# Visualization (for Voronoi analysis)
matplotlib>=3.7.0

# Optional: GPU acceleration for colorwheel
# cupy-cuda11x  # Uncomment if you have CUDA 11.x
# cupy-cuda12x  # Uncomment if you have CUDA 12.x
